+++ src/fio/FioEngine.ts
@@ -1,6 +1,7 @@
 import { FIOSDK } from '@fioprotocol/fiosdk'
 import { BalanceResponse } from '@fioprotocol/fiosdk/lib/entities/BalanceResponse'
 import { EndPoint } from '@fioprotocol/fiosdk/lib/entities/EndPoint'
+import { GetEncryptKeyResponse } from '@fioprotocol/fiosdk/lib/entities/GetEncryptKeyResponse'
 import {
   GetObtData,
   PendingFioRequests,
@@ -8,6 +9,7 @@ import {
 } from '@fioprotocol/fiosdk/lib/transactions/queries'
 import { Query } from '@fioprotocol/fiosdk/lib/transactions/queries/Query'
 import { Transactions } from '@fioprotocol/fiosdk/lib/transactions/Transactions'
+import AbortController from 'abort-controller'
 import { add, div, gt, lt, max, mul, sub } from 'biggystring'
 import { asMaybe, asString, asTuple } from 'cleaners'
 import {
@@ -628,11 +630,13 @@ export class FioEngine extends CurrencyEngine<FioTools, SafeFioWalletInfo> {
         case 'getObtData':
         case 'getPendingFioRequests':
         case 'getSentFioRequests': {
-          const { endpoint, body } = params
-          res = await fioSdk.transactions.executeCall(
-            endpoint,
-            JSON.stringify(body)
-          )
+          const { baseUrl, endpoint, body } = params
+          res = await fioSdk.transactions.executeCall({
+            baseUrl,
+            endPoint: endpoint,
+            body: JSON.stringify(body),
+            signal: new AbortController().signal
+          })
           break
         }
         case 'getCurrencyBalance': {
@@ -797,9 +801,11 @@ export class FioEngine extends CurrencyEngine<FioTools, SafeFioWalletInfo> {
           async apiUrl =>
             await timeout(
               this.fioApiRequest(apiUrl, 'getCurrencyBalance', {
-                code: 'fio.token',
-                account: this.actor,
-                symbol: 'FIO'
+                body: {
+                  code: 'fio.token',
+                  account: this.actor,
+                  symbol: 'FIO'
+                }
               }),
               10000
             )
@@ -1179,6 +1185,19 @@ export class FioEngine extends CurrencyEngine<FioTools, SafeFioWalletInfo> {
     const fioPrivateKeys = asFioPrivateKeys(opts?.privateKeys)
     let isChanged = false
 
+    // TODO: v1.9 of the FIO sdk introduced a getEncyptKey prop to their
+    // checkFioRequests fn, though the SDK can still function without it.
+    // Whenever they require it or provide more information about what it even
+    // is, we should implement it.
+    const defaultGetEncryptKey = async (
+      fioAddress: string
+    ): Promise<GetEncryptKeyResponse> => {
+      // Provide a default response or handle the fallback logic here
+      return {
+        encrypt_public_key: this.walletInfo.keys.publicKey // Use the public key as a fallback
+      }
+    }
+
     const checkFioRequests = async (
       type: FioRequestTypes,
       decoder: Query<PendingFioRequests | SentFioRequests>
@@ -1203,16 +1222,26 @@ export class FioEngine extends CurrencyEngine<FioTools, SafeFioWalletInfo> {
 
     await checkFioRequests(
       'PENDING',
-      new PendingFioRequests(this.walletInfo.keys.publicKey)
+      new PendingFioRequests({
+        fioPublicKey: this.walletInfo.keys.publicKey,
+        getEncryptKey: defaultGetEncryptKey
+      })
     )
     await checkFioRequests(
       'SENT',
-      new SentFioRequests(this.walletInfo.keys.publicKey)
+      new SentFioRequests({
+        fioPublicKey: this.walletInfo.keys.publicKey,
+        getEncryptKey: defaultGetEncryptKey
+      })
     )
 
     if (isChanged) this.localDataDirty()
 
-    const obtDecoder = new GetObtData(this.walletInfo.keys.publicKey)
+    const obtDecoder = new GetObtData({
+      fioPublicKey: this.walletInfo.keys.publicKey,
+      getEncryptKey: defaultGetEncryptKey,
+      includeEncrypted: false
+    })
     const encryptedObtData = await this.fetchEncryptedObtData(
       'getObtData',
       obtDecoder
@@ -1220,9 +1249,10 @@ export class FioEngine extends CurrencyEngine<FioTools, SafeFioWalletInfo> {
     obtDecoder.privateKey = fioPrivateKeys.fioKey
     obtDecoder.publicKey = this.walletInfo.keys.publicKey
     const decryptedObtData: { obt_data_records: ObtData[] } =
-      obtDecoder.decrypt({
-        obt_data_records: encryptedObtData
-      }) ?? { obt_data_records: [] }
+      (await obtDecoder.decrypt({
+        obt_data_records: encryptedObtData,
+        more: 0
+      })) ?? { obt_data_records: [] }
 
     this.obtData = decryptedObtData.obt_data_records