+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- changed: FIO SDK upgraded to v1.9.2
 - changed: (Piratechain) Allow  `derivePublicKey` to fail in case native code isn't present
 - removed: (Zcash/Piratechain) Remove `defaultBirthdayHeight`